+++ libs/common/src/index.ts
@@ -14,3 +14,4 @@ export * from './config';
 export * from './health';
 export * from './interceptors';
 export * from './rmq';
+export * from './migrations';